refactor: dto response와 request로 구분
