+++ src/main/java/com/leets/commitatobe/domain/user/dto/response/UserSearchResponse.java
@@ -1,4 +1,4 @@
-package com.leets.commitatobe.domain.user.dto;
+package com.leets.commitatobe.domain.user.dto.response;
 
 public record UserSearchResponse(
 	Integer ranking,
