+++ src/main/java/com/leets/commitatobe/domain/user/service/UserQueryService.java
@@ -11,9 +11,9 @@ import org.springframework.transaction.annotation.Transactional;
 import com.leets.commitatobe.domain.login.service.LoginCommandService;
 import com.leets.commitatobe.domain.tier.domain.Tier;
 import com.leets.commitatobe.domain.user.domain.User;
-import com.leets.commitatobe.domain.user.dto.UserInfoResponse;
-import com.leets.commitatobe.domain.user.dto.UserRankResponse;
-import com.leets.commitatobe.domain.user.dto.UserSearchResponse;
+import com.leets.commitatobe.domain.user.dto.response.UserInfoResponse;
+import com.leets.commitatobe.domain.user.dto.response.UserRankResponse;
+import com.leets.commitatobe.domain.user.dto.response.UserSearchResponse;
 import com.leets.commitatobe.domain.user.repository.UserRepository;
 import com.leets.commitatobe.global.exception.ApiException;
 import com.leets.commitatobe.global.response.CustomPageResponse;
