+++ src/main/java/com/leets/commitatobe/global/response/code/BaseCode.java
@@ -1,5 +1,7 @@
 package com.leets.commitatobe.global.response.code;
 
+import com.leets.commitatobe.global.response.code.dto.ReasonDto;
+
 public interface BaseCode {
 
 	public ReasonDto getReason();
