+++ src/main/java/com/leets/commitatobe/global/response/code/BaseErrorCode.java
@@ -1,5 +1,7 @@
 package com.leets.commitatobe.global.response.code;
 
+import com.leets.commitatobe.global.response.code.dto.ErrorReasonDto;
+
 public interface BaseErrorCode {
 
 	public ErrorReasonDto getReason();
